Refatoring form pages to components
