+++ src/components/Create/style.ts
@@ -7,34 +7,13 @@ export const Container = styled.section`
     min-height: calc(100vh - 100px);
     background: rgba(0,0,0,0.1);
 
-    form{
-        min-width: 500px;
-    }
-
-    label {
-        margin: 0 auto;
-        border: none;
-        font-size: 1.2rem;
-        font-weight: 600;
-        margin-bottom: 1rem;
-    }
-
-    input{
-        color:#000000;
-        margin: 0 1rem;
-        border: none;
-        width: 90%;
+    h2{
+        font-size: 2rem;
+        text-transform: uppercase;
+        display: inline-block;
+        margin-bottom: 2rem;
         border-bottom: 1px solid #CA4F2F;
-        font-size: 1rem;
-        font-weight: 500;
-        margin: .5rem 0;
-        padding: 1rem 2rem;
-
-        &:focus{
-            box-shadow: 0 0 0 0;
-            outline: 0;
-        }
-
     }
+
 `
